requirements.txt
flask
flask-cors
joblib
numpy
Pillow
scikit-learn
